+++ test/base/runner/version_test.dart
@@ -8,7 +8,7 @@ import 'package:collection/collection.dart' show ListEquality;
 //import 'package:flutter_tools/src/base/time.dart';
 import 'package:mockito/mockito.dart';
 import 'package:process/process.dart';
-import 'package:sylph/src/base/cache.dart';
+//import 'package:sylph/src/base/cache.dart';
 import 'package:sylph/src/base/version.dart';
 import 'package:test/test.dart';
 import 'package:tool_base/tool_base.dart';
